# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Lists companies",
                "operationId": "listCompanies",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyPagingResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Adds a new company",
                "operationId": "storeCompany",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/companies/{id}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get Company information",
                "description": "Returns Company data",
                "operationId": "showCompany",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Companies"
                ],
                "summary": "Update Company information",
                "description": "Returns Company data",
                "operationId": "updateCompany",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompanyUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Companies"
                ],
                "summary": "Delete Company",
                "description": "Returns deleted Company data",
                "operationId": "destroyCompany",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BaseModel": {
                "title": "BaseModel",
                "description": "BaseModel",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BaseModel"
                }
            },
            "Company": {
                "title": "Company",
                "description": "Company model",
                "type": "object",
                "xml": {
                    "name": "Company"
                },
                "allOf": [
                    {
                        "$ref": "#/components/schemas/NamedBaseModel"
                    },
                    {
                        "properties": {
                            "tenant_id": {
                                "title": "Tenant ID",
                                "description": "Tenant's id of the Company",
                                "type": "integer",
                                "format": "int64",
                                "example": 1
                            },
                            "tenant": {
                                "$ref": "#/components/schemas/Tenant"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "NamedBaseModel": {
                "title": "NamedBaseModel",
                "description": "NamedBaseModel",
                "type": "object",
                "xml": {
                    "name": "NamedBaseModel"
                },
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseModel"
                    },
                    {
                        "properties": {
                            "name": {
                                "title": "Name",
                                "description": "Name of the entity",
                                "type": "string",
                                "example": "Elmofuchto"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "Tenant": {
                "title": "Tenant",
                "description": "Tenant model",
                "xml": {
                    "name": "Tenant"
                },
                "allOf": [
                    {
                        "$ref": "#/components/schemas/NamedBaseModel"
                    }
                ]
            },
...
            "CompanyStoreRequest": {
                "title": "CompanyStoreRequest",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new company",
                        "type": "string",
                        "example": "A nice company"
                    },
                    "tenant_id": {
                        "title": "tenant_id",
                        "description": "The tenant id of the company",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "CompanyUpdateRequest": {
                "title": "CompanyUpdateRequest",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the company",
                        "type": "string",
                        "example": "A nice Tenant"
                    }
                },
                "type": "object"
            },
...
            "CompanyPagingResource": {
                "title": "CompanyPagingResource",
                "description": "CompanyPagingResource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Company"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/Meta"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CompanyPagingResource"
                }
            },
            "CompanyResource": {
                "title": "CompanyResource",
                "description": "CompanyResource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Company"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CompanyResource"
                }
            },
            "Meta": {
                "title": "Meta",
                "properties": {
                    "current_page": {
                        "title": "Current Page",
                        "description": "Current Page of the data",
                        "type": "integer"
                    },
                    "from": {
                        "title": "From",
                        "description": "From",
                        "type": "integer"
                    },
                    "last_page": {
                        "title": "Last Page",
                        "description": "Last Page of the data",
                        "type": "integer"
                    },
                    "per_page": {
                        "title": "Per Page",
                        "description": "Per Page of the data",
                        "type": "integer"
                    },
                    "to": {
                        "title": "To",
                        "description": "To",
                        "type": "integer"
                    },
                    "total": {
                        "title": "Total",
                        "description": "Total",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format Bearer 4|oeXad4kChJT43wli90LOd1VbFhtuGuEdvxvEHMAtcb025185",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}